+++ packages/zowe-explorer-api/src/vscode/ui/WebView.ts
@@ -36,14 +36,20 @@ export class WebView {
 
     /**
      * Constructs a webview for use with bundled assets.
-     * The webview entrypoint must be located at <webview folder>/dist/assets/index.js.
+     * The webview entrypoint must be located at src/<webview folder>/dist/<webview-name>/index.js.
      *
      * @param title The title for the new webview
-     * @param dirName The directory name (relative to the "webviews" folder in the extension root) with a valid entrypoint (see above).
+     * @param webviewName The webview name, the same name given to the directory of your webview in the webviews/src directory.
      * @param context The VSCode extension context
      * @param onDidReceiveMessage Event callback: called when messages are received from the webview
      */
-    public constructor(title: string, dirName: string, context: ExtensionContext, onDidReceiveMessage?: (message: object) => void | Promise<void>) {
+    public constructor(
+        title: string,
+        webviewName: string,
+        context: ExtensionContext,
+        onDidReceiveMessage?: (message: object) => void | Promise<void>,
+        retainContext?: boolean
+    ) {
         this.disposables = [];
 
         // Generate random nonce for loading the bundled script
@@ -52,13 +58,14 @@ export class WebView {
 
         // Build URIs for the webview directory and get the paths as VScode resources
         this.uris.disk = {
-            build: Uri.file(joinPath(context.extensionPath, "webviews", dirName)),
-            script: Uri.file(joinPath(context.extensionPath, "webviews", dirName, "dist", "assets", "index.js")),
+            build: Uri.file(joinPath(context.extensionPath, "src", "webviews")),
+            script: Uri.file(joinPath(context.extensionPath, "src", "webviews", "dist", webviewName, `${webviewName}.js`)),
         };
 
         this.panel = window.createWebviewPanel("ZEAPIWebview", this.title, ViewColumn.Beside, {
             enableScripts: true,
             localResourceRoots: [this.uris.disk.build],
+            retainContextWhenHidden: retainContext ?? false,
         });
 
         // Associate URI resources with webview