+++ packages/zowe-explorer/__tests__/__unit__/generators/icons.unit.test.ts
@@ -20,7 +20,7 @@ import { ZoweLocalStorage } from "../../../src/utils/ZoweLocalStorage";
 
 describe("Checking icon generator's basics", () => {
     const setGlobalMocks = () => {
-        const createTreeView = jest.fn();
+        const createTreeView = jest.fn().mockReturnValue({ onDidCollapseElement: jest.fn() });
         const getConfiguration = jest.fn();
 
         Object.defineProperty(vscode.window, "createTreeView", { value: createTreeView });