+++ packages/zowe-explorer/src/dataset/utils.ts
@@ -10,8 +10,25 @@
  */
 
 import * as globals from "../globals";
+import * as nls from "vscode-nls";
 import { IZoweNodeType } from "@zowe/zowe-explorer-api";
 import { ZoweLogger } from "../utils/LoggerUtils";
+
+// Set up localization
+nls.config({
+    messageFormat: nls.MessageFormat.bundle,
+    bundleFormat: nls.BundleFormat.standalone,
+})();
+const localize: nls.LocalizeFunc = nls.loadMessageBundle();
+
+export const DATASET_SORT_OPTS = [
+    localize("ds.sortByName", "$(case-sensitive) Name (default)"),
+    localize("ds.sortByModified", "$(calendar) Date Modified"),
+    localize("ds.sortByUserId", "$(account) User ID"),
+    localize("setSortDirection", "$(fold) Sort Direction"),
+];
+
+export const DATASET_FILTER_OPTS = [localize("ds.sortByModified", "$(calendar) Date Modified"), localize("ds.sortByUserId", "$(account) User ID")];
 
 export function getProfileAndDataSetName(node: IZoweNodeType): {
     profileName: string;